+++ course7/homeworks/mathias/chat/src/client/CMakeLists.txt
@@ -1,0 +1,3 @@
+add_executable(client client.cpp)
+set_property(TARGET client PROPERTY CXX_STANDARD 14)
+target_link_libraries(client PRIVATE procotol)
