+++ course7/homeworks/mathias/chat/src/json/CMakeLists.txt
@@ -1,0 +1,2 @@
+add_library(json INTERFACE)
+target_include_directories(json INTERFACE ${CMAKE_CURRENT_SOURCE_DIR})
