+++ course7/homeworks/mathias/chat/src/protocol/CMakeLists.txt
@@ -1,0 +1,4 @@
+add_library(procotol STATIC protocol.h protocol.cpp)
+target_link_libraries(procotol PUBLIC json Boost::system Boost::date_time)
+target_include_directories(procotol PUBLIC Boost::boost)
+target_include_directories(procotol INTERFACE ${CMAKE_CURRENT_SOURCE_DIR})
