+++ course7/homeworks/mathias/chat/src/server/CMakeLists.txt
@@ -1,0 +1,3 @@
+add_executable(server server.cpp server.h user_session.cpp user_session.h chat_room.cpp chat_room.h)
+set_property(TARGET server PROPERTY CXX_STANDARD 14)
+target_link_libraries(server PRIVATE procotol)
