+++ course7/homeworks/mathias/chat/src/server/server.h
@@ -1,0 +1,24 @@
+#include "json.hpp"
+#include "protocol.h"
+
+#include <boost/asio/bind_executor.hpp>
+#include <boost/asio/strand.hpp>
+#include <boost/asio/ip/tcp.hpp>
+#include <boost/beast/core.hpp>
+#include <boost/beast/websocket.hpp>
+#include <boost/variant/static_visitor.hpp>
+
+#include <algorithm>
+#include <cstdlib>
+#include <functional>
+#include <iostream>
+#include <memory>
+#include <string>
+#include <thread>
+#include <vector>
+
+using tcp = boost::asio::ip::tcp;
+namespace websocket = boost::beast::websocket;
+using json = nlohmann::json;
+
+
